Move Frame to own files
