+++ Bridge/src/protocol.hpp
@@ -1,38 +1,19 @@
 #pragma once
 
 #include <optional>
-#include <ostream>
 #include <stdint.h>
 #include <vector>
 
 #include <mqtt/message.h>
 
+#include "frame.hpp"
 #include "mqtt.hpp"
 #include "ring_buffer.hpp"
-
-constexpr uint16_t BODY_LENGTH = 75;
-constexpr uint16_t PACKET_LENGTH = 80;
 
 constexpr uint8_t QOS_MASK    = 0b0000'0011;
 constexpr uint8_t RETAIN_MASK = 0b0000'0100;
 
 constexpr size_t MESSAGE_HISTORY = 50;
-
-enum class FrameType : uint8_t
-{
-    Message = 0,
-};
-
-struct __attribute__((packed)) Frame
-{
-    uint8_t frame_counter;
-    FrameType frame_type;
-    uint8_t part_counter;
-    uint16_t body_length;
-    uint8_t body[BODY_LENGTH] = {0};
-};
-
-std::ostream& operator<<(std::ostream& os, const Frame *frame);
 
 class RxProtocol
 {
